hunt-commander/requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Data processing
pandas==2.1.3
python-dotenv==1.0.0
pyyaml==6.0.1

# AI/LLM integration (for CV tailoring and matching)
openai==1.3.7
anthropic==0.7.7

# Notifications
email-validator==2.1.0

# Scheduling
schedule==1.2.0

# Logging
loguru==0.7.2

# Document processing
python-docx==1.1.0
pypdf2==3.0.1
pdfkit==1.0.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Payments
stripe==7.4.0
